+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 24-06-20
+
+- roc_auc function returns direction 
+
 ## 24-05-31
 
 - ion$auc return > 0.5, default no plot.